+++ ReleaseNotes.md
@@ -1,5 +1,11 @@
 # Changelog Tiger Test platform
 
+# Release 4.0.8
+
+## Bugfixes
+
+* TGR-1897: Tiger-Proxy: Fixed parsing of chunked SICCT messages.
+
 # Release 4.0.7
 
 ## Bugfixes
@@ -814,8 +820,8 @@ import org.junit.platform.suite.api.Suite;
 @ConfigurationParameter(key = FILTER_TAGS_PROPERTY_NAME, value = "not @Ignore")
 @ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "de.gematik.test.tiger.glue,ANY ADDITIONAL PACKAGES containing GLUE or HOOKS code")
 @ConfigurationParameter(
-        key = PLUGIN_PROPERTY_NAME,
-        value = "io.cucumber.core.plugin.TigerSerenityReporterPlugin,json:target/cucumber-parallel/1.json")
+  key = PLUGIN_PROPERTY_NAME,
+  value = "io.cucumber.core.plugin.TigerSerenityReporterPlugin,json:target/cucumber-parallel/1.json")
 public class Driver1IT {
 
 }